finished refactoring components and moved current prompt from state in DAPA to Promptlayout
